Updated all UserController methods to be API Calls

+++ src/main/java/Controllers/UsersController.java
@@ -1,30 +1,47 @@
 package Controllers;
 
 import Server.Main;
+import org.glassfish.jersey.media.multipart.FormDataParam;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
+import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 @Path("UserController/")
 public class UsersController{
-
-        public static void InsertUsers(String firstName, String lastName, String password, String email, boolean admin) {
+        @POST
+        @Path("InsertUser")
+        @Consumes(MediaType.MULTIPART_FORM_DATA)
+        @Produces(MediaType.APPLICATION_JSON)
+        public String InsertUser(
+                @FormDataParam("userID") Integer userID,
+                @FormDataParam("firstName") String firstName,
+                @FormDataParam("lastName") String lastName,
+                @FormDataParam("password") String password,
+                @FormDataParam("email") String email,
+                @FormDataParam("admin") Boolean admin) {
             try{
-                PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Users (firstName, lastName, password, email, admin) values (?, ?, ?, ?, ?)");
-                ps.setString(1, firstName);
-                ps.setString(2, lastName);
-                ps.setString(3, password);
-                ps.setString(4, email);
-                ps.setBoolean(5, admin);
-                ps.executeUpdate();
-                System.out.println("Records successfully added");
+                if(userID == null || firstName == null || lastName == null || password == null || email == null || admin == null){
+                    throw new Exception("One or more form data parameters are missing from the HTTP request");
+                }
+                System.out.println("InsertUser/new userID=" + userID);
+
+                PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Users (userID, firstName, lastName, password, email, admin) values (?, ?, ?, ?, ?, ?)");
+
+                ps.setInt(1, userID);
+                ps.setString(2, firstName);
+                ps.setString(3, lastName);
+                ps.setString(4, password);
+                ps.setString(5, email);
+                ps.setBoolean(6, admin);
+                ps.execute();
+                return "{\"status\": \"OK\"}";
+
             }   catch (Exception exception){
-                System.out.println("Error: " + exception.getMessage());
+                System.out.println("Database error: " + exception.getMessage());
+                return "{\"error\": \"Unable to create new item, please see server console for more info.\"}";
             }
         }
         @GET
@@ -53,39 +70,73 @@ public class UsersController{
                 return "{\"error\": \"Unable to list items, please see server console for more info.\"}";
             }
         }
-    public static void ListUsers(int userID){
-        try{
-            PreparedStatement ps = Main.db.prepareStatement("SELECT * FROM Users WHERE userID = ?");
-            ps.setInt(1, userID);
-            ResultSet results = ps.executeQuery();
-            while(results.next()){
-                int UserID = results.getInt(1);
-                String firstName = results.getString(2);
-                String lastName = results.getString(3);
-                String password = results.getString(4);
-                String email = results.getString(5);
-                boolean admin = results.getBoolean(6);
-                System.out.println(UserID + " " + firstName + " " + lastName + " " + password + " " + email + " " + admin);
+        @GET
+        @Path("ListUser/{userID}")
+        @Produces(MediaType.APPLICATION_JSON)
+        public String ListUser(@PathParam("userID") Integer userID) throws Exception {
+            if(userID == null){
+                throw new Exception("User ID is missing from HTTP request");
             }
-        }   catch (Exception exception){
-            System.out.println("Error " + exception.getMessage());
+            System.out.println("UserController/ListUser" + userID);
+            JSONObject item = new JSONObject();
+            try{
+                PreparedStatement ps = Main.db.prepareStatement("SELECT firstName, lastName, password, email, admin FROM Users WHERE userID = ?");
+                ps.setInt(1, userID);
+                ResultSet results = ps.executeQuery();
+                if(results.next()){
+                    item.put("userID", userID);
+                    item.put("firstName", results.getString(1));
+                    item.put("lastName", results.getString(2));
+                    item.put("password", results.getString(3));
+                    item.put("email", results.getString(4));
+                    item.put("admin", results.getBoolean(5));
+                }
+                return item.toString();
+            }   catch (Exception exception){
+                System.out.println("Database error " + exception.getMessage());
+                return "{\"error\": \"Unable to get item, please see server console for more info.\"}";
+            }
         }
-    }
-        public static void DeleteUser(int userID){
+        @POST
+        @Path("DeleteUser")
+        @Consumes(MediaType.MULTIPART_FORM_DATA)
+        @Produces(MediaType.APPLICATION_JSON)
+        public String DeleteUser(@FormDataParam("userID") Integer userID){
             try {
-                PreparedStatement ps1 = Main.db.prepareStatement("DELETE FROM personalBookings WHERE userID = ?");
-                ps1.setInt(1, userID);
-                ps1.executeUpdate();
+                if(userID == null){
+                    throw new Exception("One or more form data parameters are missing in the HTTP request");
+                }
+                System.out.println("UserController/DeleteUser userID=" + userID);
+
+                //PreparedStatement ps1 = Main.db.prepareStatement("DELETE FROM personalBookings WHERE userID = ?");
+                //ps1.setInt(1, userID);
+                //ps1.execute();
                 PreparedStatement ps2 = Main.db.prepareStatement("DELETE FROM Users WHERE userID = ?");
                 ps2.setInt(1, userID);
-                ps2.executeUpdate();
-                System.out.println("Records removed successfully");
+                ps2.execute();
+                return "{\"status\": \"OK\"}";
+
             } catch (Exception e){
-                System.out.println("Error " + e.getMessage());
+                System.out.println("Database error: " + e.getMessage());
+                return "{\"error\": \"Unable to delete item, please see server console for more info.\"}";
             }
         }
-        public static void UpdateUser(int userID, String firstName, String lastName, String password, String email, boolean admin){
+        @POST
+        @Path("UpdateUser")
+        @Consumes(MediaType.MULTIPART_FORM_DATA)
+        @Produces(MediaType.APPLICATION_JSON)
+        public String UpdateUser(
+                @FormDataParam("userID") Integer userID,
+                @FormDataParam("firstName") String firstName,
+                @FormDataParam("lastName") String lastName,
+                @FormDataParam("password") String password,
+                @FormDataParam("email") String email,
+                @FormDataParam("admin") Boolean admin) {
             try {
+                if(userID == null || firstName == null || lastName == null || password == null || email == null || admin == null){
+                    throw new Exception("One or more form data parameters are missing in the HTTP request");
+                }
+                System.out.println("UserController/UpdateUser userID=" + userID);
                 PreparedStatement ps = Main.db.prepareStatement("UPDATE Users SET firstName = ?, lastName = ?, password = ?, email = ?, admin = ? WHERE userID = ?");
                 ps.setString(1, firstName);
                 ps.setString(2, lastName);
@@ -93,10 +144,12 @@ public class UsersController{
                 ps.setString(4, email);
                 ps.setBoolean(5, admin);
                 ps.setInt(6, userID);
-                ps.executeUpdate();
-                System.out.println("Records successfully updated");
+                ps.execute();
+                return "{\"status\": \"OK\"}";
+
             }catch (Exception e){
-                System.out.println("Error " + e.getMessage());
+                System.out.println("Database error: " + e.getMessage());
+                return "{\"error\": \"Unable to update item, please see server console for more info\"}";
             }
         }
